+++ lib/ext/Explore.dart
@@ -2,6 +2,7 @@ import 'dart:ui';
 
 import 'package:geolocator/geolocator.dart';
 import 'package:illinois/model/Dining.dart';
+import 'package:illinois/model/Explore.dart';
 import 'package:illinois/model/Laundry.dart';
 import 'package:illinois/model/MTD.dart';
 import 'package:illinois/model/StudentCourse.dart';
@@ -126,6 +127,9 @@ extension ExploreExt on Explore {
     else if (exploresType == "appointment") {
       return Localization().getStringEx('panel.explore.item.appointments.name', 'Appointments');
     }
+    else if (exploresType == "explorepoi") {
+      return Localization().getStringEx('panel.explore.item.pois.name', 'POIs');
+    }
     else {
       return Localization().getStringEx('panel.explore.item.unknown.name', 'Explores');
     }
@@ -198,6 +202,9 @@ extension ExploreExt on Explore {
     else if (this is StudentCourse) {
       return (this as StudentCourse).uiColor;
     }
+    else if (this is ExplorePOI) {
+      return (this as ExplorePOI).uiColor;
+    }
     //else if (this is Building) {}
     //else if (this is Appointment) {}
     else {
@@ -206,4 +213,8 @@ extension ExploreExt on Explore {
   }
 
   String? get exploreImageUrl => (this is Event) ? (this as Event).eventImageUrl : exploreImageURL;
+}
+
+extension ExplorePOIExt on ExplorePOI {
+  Color? get uiColor => Styles().colors?.accentColor3;
 }